UserServiceImpl

+++ backend/src/main/java/net/bcsoft/careergraph/dto/UserDTO.java
@@ -2,7 +2,7 @@ package net.bcsoft.careergraph.dto;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 
-public record UserDTO(Integer id, @JsonProperty("sso_uid") String ssoUid, @JsonProperty("first_name") String firstName, @JsonProperty("last_name") String lastName, String email) {}
+public record UserDTO(Long id, @JsonProperty("sso_uid") String ssoUid, @JsonProperty("first_name") String firstName, @JsonProperty("last_name") String lastName, String email) {}
 
 /*
 public class UserDTO {
